Desarrollo Backend con Python/Curso de FastAPI - Base de Datos, Modularización y Deploy a Producción/scripts/my_movie_api with sqlmodel/requirements.txt
fastapi==0.89.1
pydantic==1.9.2
PyJWT==2.6.0
sqlmodel==0.0.8
